added new socket data type that works with both monochromatic and color images

+++ Logic/Nodes/HomographyNodes.cpp
@@ -127,7 +127,7 @@ public:
 			{ ENodeFlowDataType::Invalid, "", "", "" }
 		};
 		static const PropertyConfig prop_config[] = {
-			{ EPropertyType::Integer, "Reprojection error threshold", "min:1.0, max:50.0" },
+			{ EPropertyType::Double, "Reprojection error threshold", "min:1.0, max:50.0" },
 			{ EPropertyType::Unknown, "", "" }
 		};
